Move bot court
